Split lint functions into Lint() and LintFiles()
Lint() lints a single file and takes an io.Reader and filename. This
takes over the responsibility of the previous parse() function and
provides a way for us to easily handle stdin-based linting in the
future.

LintFiles() is equivalent to the previous Lint() function. It takes a
list of filenames. Each is opened, parsed, and linted in turn, and the
combined results are returned.