Reimplement overridableChecks using a list
Using a map gave good runtime performance, but we ran into one of the
disadvantages of using interface keys: all types that implement the
interface needs to support map key semantics.

Unfortunately, some of the node types (like MapType and SetType) are not
hashable, and that broke this implementation.

This change uses a list of overrides instead of a map, which works well
and isn't much more complex in practice.