Use filepath.Match() to match keys in include.restricted (#9)
The previous regular expression syntax was a bit heavy for the left-hand
side of this table. We can instead use the simpler file name pattern
syntax.

I'm keeping the richer regular expression syntax for the right-hand side
because it gives us a nice way to match multiple unrelated files using
groups. The alternative would be to express the configuration block as a
list of (pattern, pattern) pairs (rather than a map), which would allow
repeated left-hand side values, but that doesn't feel like the right
trade-off.