Use fnmatch()-style file path matching (#18)
filepath.Match() is aware of directory separators. This makes sense for
shell-style matching but isn't as convenient working with arbitrary file
paths because the directory hierarchy needs to be expressed in the
pattern.

fnmatch(3) offers a FNM_NOESCAPE flag that avoid that problem. When set,
directory separators are treated like normal characters. This is more
intuitive for our use case (and matches Python's fnmatch.fnmatch()).

We can use github.com/danwakefield/fnmatch for a Go implementation of
the fnmatch(3) matching rules.