Use current directory to resolve includes (#24)
We were previously using the including file's name, which I mistakenly
thought was the `thrift` compiler's behavior. We instead need to add the
current directory as the first search path to match `thrift`.

This change also adds support for absolute paths and makes consistent
use of the `path/filepath` package for better cross-platform support.