Fix include search path logic
This reverts part of the changes introduced in #24: we want to include
the file's directory in our search path, not the process's current
working directory.

In fixing this, I realized that we can simply build the list of search
directories once as part of the linting context. This further simplifies
the downstream code while also making this less error prone.