Make all types-based checks consistent
All type-based checks (maps, sets, and "global") now use consistent
logic and configuration keys.

They can be configured with "allowedTypes" and "disallowedTypes" lists.
If the type appears in "disallowedTypes", an error is reported. If the
list "allowedTypes" is not empty, then the type must appear in it, or an
error is reported.

This is a breaking change for some existing checks, but it gives us a
lot more flexibility after a one-time configuration update, so lets roll
with it.